Swap badge for static version number
The badge is an external resource and adds an extra request to the page.
It is also inaccessible because the version number is within a `img`,
where assistive technology cannot access it.

This commit uses the `middleman-data_source` gem to mount remote JSON
as local data from the RubyGems API, which we can pull the latest
version number from.

See: https://github.com/stevenosloan/middleman-data_source